+++ src/org/labkey/remoteapi/CredentialsProvider.java
@@ -17,11 +17,20 @@ package org.labkey.remoteapi;
 
 import org.apache.hc.client5.http.auth.AuthenticationException;
 import org.apache.hc.client5.http.classic.methods.HttpUriRequest;
+import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
 import org.apache.hc.client5.http.protocol.HttpClientContext;
 
+import java.io.IOException;
 import java.net.URI;
 
 public interface CredentialsProvider
 {
+    void configureClientBuilder(URI baseURI, HttpClientBuilder builder);
     void configureRequest(URI baseURI, HttpUriRequest request, HttpClientContext httpClientContext) throws AuthenticationException;
+
+    /**
+     * Initialize the connection before its first request. If connection-based, authenticate the user. In all cases,
+     * retrieve the CSRF token and session ID to use with subsequent requests on this connection.
+     */
+    void initializeConnection(Connection connection) throws IOException, CommandException;
 }